+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## unreleased 
 * Remove support for CardIO, a no longer maintained framework. (NEXT_MAJOR_VERSION)
+* Change postal code default keyboard type to `UIKeyboardTypeDefault` (NEXT_MAJOR_VERSION)
+* Update podspec to require versions `~> 4.31` of Braintree SDK dependencies.
 
 ## 7.5.0 (2019-11-19)