+++ CHANGELOG.md
@@ -6,6 +6,9 @@
   * Remove BTDropInBaseViewController showLoadingScreen:animated: method (NEXT_MAJOR_VERSION)
   * Remove BTDropInRequest amount property (NEXT_MAJOR_VERSION)
   * Remove BTUIKAppearance darkTheme and lightTheme methods (NEXT_MAJOR_VERSION)
+* Remove support for CardIO, a no longer maintained framework. (NEXT_MAJOR_VERSION)
+* Change postal code default keyboard type to `UIKeyboardTypeDefault` (NEXT_MAJOR_VERSION)
+* Update podspec to require versions `~> 4.31` of Braintree SDK dependencies.
 
 ## 7.5.0 (2019-11-19)