+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## unreleased
 
-* Fix several UIKit deprecation warnings
+* Fix UIKit deprecation warnings:
+  * UIApplication's `statusBarOrientation` and `statusBarFrame` properties
+  * UIViewController's `topLayoutGuide` and `bottomLayoutGuide` properties
 
 ## 8.0.0 (2020-02-06)